requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
aiohttp==3.9.1
pydantic==2.5.0
python-multipart==0.0.6
slowapi==0.1.9
redis==5.0.1
python-dotenv==1.0.0

# Background job processing
celery==5.3.4
rq==1.15.1

# Structured logging
structlog==23.2.0
python-json-logger==2.0.7

# Storage backends
motor==3.3.2  # MongoDB async driver
pymongo==4.6.0  # MongoDB sync driver
elasticsearch==8.11.0  # Elasticsearch client